requirements/base.txt
argon2-cffi==21.3.0
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.16
django-environ==0.9.0
psycopg2-binary==2.9.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
pycparser==2.21
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
